Moved general class to glue.dialogs.common
